Move renderLinks to helper
