+++ src/helper.tsx
@@ -1,0 +1,28 @@
+import Link from '@material-ui/core/Link';
+import React from 'react';
+
+export function renderLinks(markdown: string) {
+    const linkCatcher = /\[(.+?)\]\((.+?)\)/g;
+    const matches = markdown.matchAll(linkCatcher);
+
+    if (matches === null) {
+        return markdown;
+    }
+
+    let i = 0;
+    let contents: any[] = [];
+
+    for (const match of matches) {
+        contents.push(markdown.substr(i, markdown.indexOf(match[0]) - i));
+        contents.push(
+            <Link href={match[2]} color="secondary">
+                {match[1]}
+            </Link>
+        );
+        i += markdown.indexOf(match[0]) + match[0].length;
+    }
+
+    contents.push(markdown.substr(i, markdown.length));
+
+    return contents;
+}
